pyproject.toml
[tool.poetry]
name = "fastapi-template"
version = "0.1.0"
description = ""
authors = ["valon <valon.januzaj98@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
pydantic-settings = "^2.4.0"
sqlalchemy = {extras = ["mypy"], version = "^2.0.36"}
alembic = "^1.13.2"
psycopg = {version = "^3.1.13", extras = ["binary"]}
celery = {extras = ["redis"], version = "^5.4.0"}
gunicorn = "^23.0.0"
inflection = "^0.5.1"
passlib = "^1.7.4"
pydantic = {extras = ["email"], version = "^2.9.2"}
python-multipart = "^0.0.16"
pyjwt = "^2.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
httpx = "^0.27.2"
ruff = "^0.6.3"
pytest-sugar = "^1.0.0"
mypy = "^1.11.2"
pytest-celery = "^1.1.2"
types-jwt = "^0.1.3"
types-passlib = "^1.7.7.20240819"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
strict = true
exclude = ["venv", ".venv", "alembic"]
plugins = ["sqlalchemy.ext.mypy.plugin"]

[tool.ruff]
target-version = "py312"
exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "B",  # flake8-bugbear
    "W",  # pycodestyle warnings
    "I",  # isort
    "E",  # pycodestyle errors
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "F",  # pyflakes
    "ARG001", # unused arguments in functions
]
ignore = [
    "E501",  # line too long
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true
